Temporary fix for timer problem
The songInterval would stop running when the timer was active, so the currently playing song would stop updating. For some reason, the intervals only run together if the delays are the same. I'm afraid of Spotify rate limiting, so I don't want to fetch the song every second, but this is a temporary fix